refactor to Interpreter interface
I intend to continue with Go for the VM implementation of golox.

In Crafting Interpreters, C is used probably to demonstrate implementing an
entire VM with a low-level minimal language with manual memory management. But
as I am just using this as a learning tool, I want to keep the implementation
pace with a programming language I am familiar with rather than C which I have
not touched in decades.

Ref:
https://craftinginterpreters.com/a-bytecode-virtual-machine.html