pratt parser for expressions
It is kind of ugly because the original C source relies on a lot of global
variables. Since I want the code to follow the book I had to make some
concessions to incorporate this global variable style into the current code.
This means the parser global should be the same even though the parser methods
and a few others operate on a parser reference. In other words, parser
references should always be refering to the global reference.

17 Compiling Expressions
https://craftinginterpreters.com/compiling-expressions.html